@@ -1,12 +1,11 @@
 name: flutter_blue_elves
 description: A flutter plugin for bluetooth ble device connect and control.
-version: 0.0.1
-author: PineappleOilPrince
+version: 0.1.0
 homepage: https://github.com/pineappleoOilPrince/flutter_blue_elves
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.10.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
