@@ -9,11 +9,11 @@ import 'package:flutter_blue_elves/flutter_blue_elves.dart';
 import 'dart:io';
 
 class DeviceControl extends StatefulWidget {
-  final String _macAddress;
-  final String _name;
+  final String? _macAddress;
+  final String? _name;
   final Device _device;
 
-  const DeviceControl(this._name, this._macAddress, this._device, {Key key})
+  const DeviceControl(this._name, this._macAddress, this._device, {Key? key})
       : super(key: key);
 
   @override
@@ -23,11 +23,11 @@ class DeviceControl extends StatefulWidget {
 class _DeviceControlState extends State<DeviceControl> {
   final List<_ServiceListItem> _serviceInfos;
   final TextEditingController _sendDataTextController = TextEditingController();
-  DeviceState _deviceState;
+  late DeviceState _deviceState;
   final List<_LogItem> _logs = [];
-  StreamSubscription<BleService> _serviceDiscoveryStream;
-  StreamSubscription<DeviceState> _stateStream;
-  StreamSubscription<DeviceSignalResult> _deviceSignalResultStream;
+  late StreamSubscription<BleService> _serviceDiscoveryStream;
+  late StreamSubscription<DeviceState> _stateStream;
+  late StreamSubscription<DeviceSignalResult> _deviceSignalResultStream;
 
   _DeviceControlState(this._serviceInfos);
 
@@ -68,18 +68,19 @@ class _DeviceControlState extends State<DeviceControl> {
     });
     _deviceSignalResultStream =
         widget._device.deviceSignalResultStream.listen((event) {
-      if (event.type == DeviceSignalType.characteristicsRead||event.type == DeviceSignalType.unKnown) {
-        String data;
-        if (event.data != null && event.data.isNotEmpty) {
-          data = "0x";
-          for (int i = 0; i < event.data.length; i++) {
-            String currentStr = event.data[i].toRadixString(16).toUpperCase();
-            if (currentStr.length < 2) {
-              currentStr = "0" + currentStr;
-            }
-            data = data + currentStr;
+      String? data;
+      if (event.data != null && event.data!.isNotEmpty) {
+        data = "0x";
+        for (int i = 0; i < event.data!.length; i++) {
+          String currentStr = event.data![i].toRadixString(16).toUpperCase();
+          if (currentStr.length < 2) {
+            currentStr = "0" + currentStr;
           }
+          data = data! + currentStr;
         }
+      }
+      if (event.type == DeviceSignalType.characteristicsRead ||
+          event.type == DeviceSignalType.unKnown) {
         setState(() {
           _logs.insert(
               0,
@@ -92,17 +93,6 @@ class _DeviceControlState extends State<DeviceControl> {
                   DateTime.now().toString()));
         });
       } else if (event.type == DeviceSignalType.characteristicsWrite) {
-        String data;
-        if (event.data != null && event.data.isNotEmpty) {
-          data = "0x";
-          for (int i = 0; i < event.data.length; i++) {
-            String currentStr = event.data[i].toRadixString(16).toUpperCase();
-            if (currentStr.length < 2) {
-              currentStr = "0" + currentStr;
-            }
-            data = data + currentStr;
-          }
-        }
         setState(() {
           _logs.insert(
               0,
@@ -115,36 +105,19 @@ class _DeviceControlState extends State<DeviceControl> {
                   DateTime.now().toString()));
         });
       } else if (event.type == DeviceSignalType.characteristicsNotify) {
-        String data = "0x";
-        for (int i = 0; i < event.data.length; i++) {
-          String currentStr = event.data[i].toRadixString(16).toUpperCase();
-          if (currentStr.length < 2) {
-            currentStr = "0" + currentStr;
-          }
-          data += currentStr;
-        }
         setState(() {
-          _logs.insert(
-              0, _LogItem(event.uuid, data, DateTime.now().toString()));
+          _logs.insert(0,
+              _LogItem(event.uuid, data ?? "none", DateTime.now().toString()));
         });
       } else if (event.type == DeviceSignalType.descriptorRead) {
-        String data;
-        if (event.data != null && event.data.isNotEmpty) {
-          data = "0x";
-          for (int i = 0; i < event.data.length; i++) {
-            String currentStr = event.data[i].toRadixString(16).toUpperCase();
-            if (currentStr.length < 2) {
-              currentStr = "0" + currentStr;
-            }
-            data = data + currentStr;
-          }
-        }
         setState(() {
           _logs.insert(
               0,
               _LogItem(
                   event.uuid,
-                  (event.isSuccess ? "read descriptor data success signal and data:" : "read descriptor data failed signal and data:") +
+                  (event.isSuccess
+                          ? "read descriptor data success signal and data:"
+                          : "read descriptor data failed signal and data:") +
                       (data ?? "none"),
                   DateTime.now().toString()));
         });
@@ -161,9 +134,11 @@ class _DeviceControlState extends State<DeviceControl> {
           Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
-             FlatButton(
+              TextButton(
                 child: Text(
-                    _deviceState == DeviceState.connected ? "Disconnect" : "connect",
+                    _deviceState == DeviceState.connected
+                        ? "Disconnect"
+                        : "connect",
                     style: const TextStyle(color: Colors.white)),
                 onPressed: () {
                   if (_deviceState == DeviceState.connected) {
@@ -206,7 +181,8 @@ class _DeviceControlState extends State<DeviceControl> {
             },
             children: _serviceInfos.map((service) {
               String serviceTitle = "Unknow Service";
-              if(Platform.isAndroid){//安卓才能发现到这些服务
+              if (Platform.isAndroid) {
+                //安卓才能发现到这些服务
                 switch (service._serviceInfo.serviceUuid.substring(4, 8)) {
                   case "1800":
                     serviceTitle = "Generic Access";
@@ -243,10 +219,10 @@ class _DeviceControlState extends State<DeviceControl> {
                     children: service._serviceInfo.characteristics
                         .map((characteristic) {
                       String properties = "";
-                      List<RaisedButton> buttons = [];
+                      List<ElevatedButton> buttons = [];
                       if (characteristic.properties
                           .contains(CharacteristicProperties.read)) {
-                        buttons.add(RaisedButton(
+                        buttons.add(ElevatedButton(
                           onPressed: () {
                             widget._device.readData(
                                 service._serviceInfo.serviceUuid,
@@ -259,7 +235,7 @@ class _DeviceControlState extends State<DeviceControl> {
                               .contains(CharacteristicProperties.write) ||
                           characteristic.properties.contains(
                               CharacteristicProperties.writeNoResponse)) {
-                        buttons.add(RaisedButton(
+                        buttons.add(ElevatedButton(
                           child: const Text("Write"),
                           onPressed: () {
                             showDialog<void>(
@@ -272,10 +248,11 @@ class _DeviceControlState extends State<DeviceControl> {
                                       autofocus: true,
                                       controller: _sendDataTextController,
                                       decoration: const InputDecoration(
-                                        hintText: "Enter hexadecimal,such as FED10101",
+                                        hintText:
+                                            "Enter hexadecimal,such as FED10101",
                                       ),
                                     ),
-                                    FlatButton(
+                                    TextButton(
                                       child: const Text("Send"),
                                       onPressed: () {
                                         String dataStr =
@@ -310,7 +287,7 @@ class _DeviceControlState extends State<DeviceControl> {
                               .contains(CharacteristicProperties.notify) ||
                           characteristic.properties
                               .contains(CharacteristicProperties.indicate)) {
-                        buttons.add(RaisedButton(
+                        buttons.add(ElevatedButton(
                           child: const Text("Set Notify"),
                           onPressed: () {
                             showDialog<void>(
@@ -425,8 +402,10 @@ class _DeviceControlState extends State<DeviceControl> {
                                                     .map((descriptor) {
                                                   String descriptorType =
                                                       "UnKnown";
-                                                  switch (Platform.isAndroid?descriptor.uuid
-                                                      .substring(4, 8):descriptor.uuid) {
+                                                  switch (Platform.isAndroid
+                                                      ? descriptor.uuid
+                                                          .substring(4, 8)
+                                                      : descriptor.uuid) {
                                                     case "2900":
                                                       descriptorType =
                                                           "Characteristic Extended Properties";
@@ -490,7 +469,7 @@ class _DeviceControlState extends State<DeviceControl> {
                                                         ],
                                                       ),
                                                       Row(children: [
-                                                        RaisedButton(
+                                                        ElevatedButton(
                                                           child: const Text(
                                                               "Read"),
                                                           onPressed: () {
@@ -506,7 +485,7 @@ class _DeviceControlState extends State<DeviceControl> {
                                                         ),
                                                         descriptorType ==
                                                                 "Client Characteristic Configuration"
-                                                            ? RaisedButton(
+                                                            ? ElevatedButton(
                                                                 child:
                                                                     const Text(
                                                                         "Write"),
@@ -533,7 +512,7 @@ class _DeviceControlState extends State<DeviceControl> {
                                                                               hintText: "Enter hexadecimal,such as FED10101",
                                                                             ),
                                                                           ),
-                                                                          FlatButton(
+                                                                          TextButton(
                                                                             child:
                                                                                 const Text("Send"),
                                                                             onPressed:
